requirements.txt
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
spacy>=3.7.0
nltk>=3.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
textblob>=0.17.0
vaderSentiment>=3.3.0
openai>=0.28.0
anthropic>=0.7.0
google-generativeai>=0.3.0
requests>=2.31.0
streamlit>=1.28.0 